+++ tests/communication/testregistervaluehandler.cpp
@@ -27,12 +27,12 @@ void TestRegisterValueHandler::cleanup()
 void TestRegisterValueHandler::addressList_16()
 {
     auto modbusRegisters = QList<ModbusRegister>() << ModbusRegister(40001, Connection::ID_1, false, true);
-    auto expRegisterList = QList<uint16_t>() << 40001;
+    auto expRegisterList = QList<quint32>() << 40001;
 
     RegisterValueHandler regHandler(_pSettingsModel);
     regHandler.setRegisters(modbusRegisters);
 
-    QList<quint16> actualRegisterList;
+    QList<quint32> actualRegisterList;
     regHandler.registerAddresList(actualRegisterList, Connection::ID_1);
 
     QVERIFY(actualRegisterList == expRegisterList);
@@ -42,12 +42,12 @@ void TestRegisterValueHandler::addressList_16_multiple()
 {
     auto modbusRegisters = QList<ModbusRegister>() << ModbusRegister(40001, Connection::ID_1, false, true)
                                                    << ModbusRegister(40002, Connection::ID_1, false, true);
-    auto expRegisterList = QList<uint16_t>() << 40001 << 40002;
+    auto expRegisterList = QList<quint32>() << 40001 << 40002;
 
     RegisterValueHandler regHandler(_pSettingsModel);
     regHandler.setRegisters(modbusRegisters);
 
-    QList<quint16> actualRegisterList;
+    QList<quint32> actualRegisterList;
     regHandler.registerAddresList(actualRegisterList, Connection::ID_1);
 
     QVERIFY(actualRegisterList == expRegisterList);
@@ -56,12 +56,12 @@ void TestRegisterValueHandler::addressList_16_multiple()
 void TestRegisterValueHandler::addressList_32()
 {
     auto modbusRegisters = QList<ModbusRegister>() << ModbusRegister(40001, Connection::ID_1, true, true);
-    auto expRegisterList = QList<uint16_t>() << 40001 << 40002;
+    auto expRegisterList = QList<quint32>() << 40001 << 40002;
 
     RegisterValueHandler regHandler(_pSettingsModel);
     regHandler.setRegisters(modbusRegisters);
 
-    QList<quint16> actualRegisterList;
+    QList<quint32> actualRegisterList;
     regHandler.registerAddresList(actualRegisterList, Connection::ID_1);
 
     QVERIFY(actualRegisterList == expRegisterList);
@@ -71,12 +71,12 @@ void TestRegisterValueHandler::addressList_32_multiple()
 {
     auto modbusRegisters = QList<ModbusRegister>() << ModbusRegister(40001, Connection::ID_1, true, true)
                                                    << ModbusRegister(40005, Connection::ID_1, true, true);
-    auto expRegisterList = QList<uint16_t>() << 40001 << 40002 << 40005 << 40006;
+    auto expRegisterList = QList<quint32>() << 40001 << 40002 << 40005 << 40006;
 
     RegisterValueHandler regHandler(_pSettingsModel);
     regHandler.setRegisters(modbusRegisters);
 
-    QList<quint16> actualRegisterList;
+    QList<quint32> actualRegisterList;
     regHandler.registerAddresList(actualRegisterList, Connection::ID_1);
 
     QVERIFY(actualRegisterList == expRegisterList);
@@ -88,12 +88,12 @@ void TestRegisterValueHandler::addressListCombinations()
                                                    << ModbusRegister(40005, Connection::ID_1, false, true)
                                                    << ModbusRegister(40005, Connection::ID_1, true, true)
                                                    << ModbusRegister(40008, Connection::ID_1, true, true);
-    auto expRegisterList = QList<uint16_t>() << 40001 << 40005 << 40006 << 40008 << 40009;
+    auto expRegisterList = QList<quint32>() << 40001 << 40005 << 40006 << 40008 << 40009;
 
     RegisterValueHandler regHandler(_pSettingsModel);
     regHandler.setRegisters(modbusRegisters);
 
-    QList<quint16> actualRegisterList;
+    QList<quint32> actualRegisterList;
     regHandler.registerAddresList(actualRegisterList, Connection::ID_1);
 
     QVERIFY(actualRegisterList == expRegisterList);
@@ -104,14 +104,14 @@ void TestRegisterValueHandler::addressListMultipleConnections()
     auto modbusRegisters = QList<ModbusRegister>() << ModbusRegister(40001, Connection::ID_1, true, true)
                                                    << ModbusRegister(40005, Connection::ID_2, true, true)
                                                    << ModbusRegister(40010, Connection::ID_1, true, true);
-    auto expRegisterList0 = QList<uint16_t>() << 40001 << 40002 << 40010 << 40011;
-    auto expRegisterList1 = QList<uint16_t>() << 40005 << 40006;
+    auto expRegisterList0 = QList<uint32_t>() << 40001 << 40002 << 40010 << 40011;
+    auto expRegisterList1 = QList<uint32_t>() << 40005 << 40006;
 
     RegisterValueHandler regHandler(_pSettingsModel);
     regHandler.setRegisters(modbusRegisters);
 
-    QList<quint16> actualRegisterList0;
-    QList<quint16> actualRegisterList1;
+    QList<quint32> actualRegisterList0;
+    QList<quint32> actualRegisterList1;
     regHandler.registerAddresList(actualRegisterList0, Connection::ID_1);
     regHandler.registerAddresList(actualRegisterList1, Connection::ID_2);
 
@@ -123,7 +123,7 @@ void TestRegisterValueHandler::read_16()
 {
     auto modbusRegisters = QList<ModbusRegister>() << ModbusRegister(40001, Connection::ID_1, false, true)
                                                    << ModbusRegister(40002, Connection::ID_1, false, false);
-    QMap<quint16, Result> partialResultMap;
+    QMap<quint32, Result> partialResultMap;
     addToResultMap(partialResultMap, 40001, false, 256, true);
     addToResultMap(partialResultMap, 40002, false, -100, true);
 
@@ -137,7 +137,7 @@ void TestRegisterValueHandler::read_32()
 {
     auto modbusRegisters = QList<ModbusRegister>() << ModbusRegister(40001, Connection::ID_1, true, true)
                                                    << ModbusRegister(40005, Connection::ID_1, true, false);
-    QMap<quint16, Result> partialResultMap;
+    QMap<quint32, Result> partialResultMap;
     addToResultMap(partialResultMap, 40001, true, 1000000, true);
     addToResultMap(partialResultMap, 40005, true, -100000, true);
 
@@ -152,7 +152,7 @@ void TestRegisterValueHandler::readBigEndian_32()
     _pSettingsModel->setInt32LittleEndian(Connection::ID_1, false);
 
     auto modbusRegisters = QList<ModbusRegister>() << ModbusRegister(40001, Connection::ID_1, true, true);
-    QMap<quint16, Result> partialResultMap;
+    QMap<quint32, Result> partialResultMap;
     quint32 value = 1000000;
 
     partialResultMap.insert(40001, Result(static_cast<quint32>(value) >> 16, true));
@@ -168,7 +168,7 @@ void TestRegisterValueHandler::readBigEndian_s32()
     _pSettingsModel->setInt32LittleEndian(Connection::ID_1, false);
 
     auto modbusRegisters = QList<ModbusRegister>() << ModbusRegister(40001, Connection::ID_1, true, true);
-    QMap<quint16, Result> partialResultMap;
+    QMap<quint32, Result> partialResultMap;
     quint32 value = -1000000;
 
     partialResultMap.insert(40001, Result(static_cast<quint32>(value) >> 16, true));
@@ -183,10 +183,10 @@ void TestRegisterValueHandler::readConnections()
 {
     auto modbusRegisters = QList<ModbusRegister>() << ModbusRegister(40001, Connection::ID_1, false, true)
                                                    << ModbusRegister(40001, Connection::ID_2, false, false);
-    QMap<quint16, Result> partialResultMap1;
+    QMap<quint32, Result> partialResultMap1;
     addToResultMap(partialResultMap1, 40001, false, 256, true);
 
-    QMap<quint16, Result> partialResultMap2;
+    QMap<quint32, Result> partialResultMap2;
     addToResultMap(partialResultMap2, 40001, false, 100, true);
 
     auto expResults = QList<Result>() << Result(256, true)
@@ -220,7 +220,7 @@ void TestRegisterValueHandler::readFail()
     auto modbusRegisters = QList<ModbusRegister>() << ModbusRegister(40001, Connection::ID_1, false, true)
                                                    << ModbusRegister(40001, Connection::ID_2, false, false);
 
-    QMap<quint16, Result> partialResultMap2;
+    QMap<quint32, Result> partialResultMap2;
     addToResultMap(partialResultMap2, 40001, false, 100, true);
 
     auto expResults = QList<Result>() << Result(0, false)
@@ -249,7 +249,7 @@ void TestRegisterValueHandler::readFail()
 }
 
 void TestRegisterValueHandler::verifyRegisterResult(QList<ModbusRegister>& regList,
-                                                    QMap<quint16, Result> &regData,
+                                                    QMap<quint32, Result> &regData,
                                                     QList<Result> expResults)
 {
     RegisterValueHandler regHandler(_pSettingsModel);
@@ -273,8 +273,8 @@ void TestRegisterValueHandler::verifyRegisterResult(QList<ModbusRegister>& regLi
     QCOMPARE(result, expResults);
 }
 
-void TestRegisterValueHandler::addToResultMap(QMap<quint16, Result> &resultMap,
-        quint16 addr,
+void TestRegisterValueHandler::addToResultMap(QMap<quint32, Result> &resultMap,
+        quint32 addr,
         bool b32bit,
         qint64 value,
         bool result
